backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-multipart
loguru
python-jose[cryptography]
python-dotenv
openpyxl
passlib[bcrypt]